OH LOOK A CIVILIAN AIRLINER
Starting CRAM implementation for tracking

Light shells are more impactful with sounds and better colors for particles

Fixed issues with rockets going r/wooooosh when not having a target

Fixed issues with explosions not doing damage

Fixed issues with missile launcher committing unalive

Changed interactions to be more proper

Crouching isn't required to stop targeting now

artillery now uses a special STICK

You can now make the flak shells from the heavy flak explode when you want it to

A lot of organization and refactoring

Followed through with the config

Artillery now has an interface for health and rotation and stuff when riding

Rapid fire guns like the light flak and
heavy machinegun can now overheat and jam (WIP)

CUSTOM EXPLOSIONS cause WHY NOT?
(Also WIP)

+++ src/main/java/me/camm/productions/fortressguns/Artillery/Entities/MultiEntityGuns/HeavyFlak.java
@@ -17,11 +17,9 @@ import org.jetbrains.annotations.NotNull;
 public class HeavyFlak extends FlakArtillery implements AutoTracking {
 
     private static final int SMALL_THRESH = 3;
-    private static int maxHealth;
+    private static double maxHealth;
     private static long fireCooldown;
 
-    //length of body before starting the barrel
-   // private static final int BODY_LENGTH = 3;
 
     public HeavyFlak(Location loc, World world, ChunkLoader loader, EulerAngle aim) {
         super(loc, world,loader,aim);
@@ -37,7 +35,7 @@ public class HeavyFlak extends FlakArtillery implements AutoTracking {
         fireCooldown = 5000;
     }
 
-    public static void setMaxHealth(int health) {
+    public static void setMaxHealth(double health) {
         maxHealth = health;
     }
